refactor common use functions to appropriate modules for better access & reuse.
